+++ uo_sys_framework/AOLimitModule/cfgFunctions.hpp
@@ -2,6 +2,7 @@ class UO_FW_AOLimitFunctions
 {
 	file = "\x\UO_FW\addons\main\AOLimitModule\cfgFunctions\common";
 	class AOLimitInit	{};
+	class DisplayAOLimitWarning {};
 };
 class UO_FW_AOLimitModuleFunctions {
 	file = "\x\UO_FW\addons\main\AOLimitModule\cfgFunctions\modules";